Move config to common module
