+++ services/api/alembic/env.py
@@ -11,7 +11,7 @@ BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 REPO_DIR = os.path.abspath(os.path.join(BASE_DIR, "../.."))
 sys.path.append(REPO_DIR)
 
-from sidetrack.api.config import get_settings  # type: ignore
+from sidetrack.common.config import get_settings  # type: ignore
 from sidetrack.common.models import Base  # type: ignore
 
 # this is the Alembic Config object, which provides
